+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.13.3 (tbd)
 
-- Implement `PartialSmt` (#372).
+- Implement `PartialSmt` (#372, #381).
 
 ## 0.13.2 (2025-01-24)